requirements.txt
fastapi
uvicorn
python-dotenv
groq
httpx
python-multipart
pytesseract
pillow
requests
httpx